Updates to new player api, kafka error handling, and general jaxrs handling
Player is moving to a new api, meaning we have to update player client
to handle the new location of the credentials within the response.

Kafka occasionally encounters oddness when running locally, that looks
like it's traceable back to dns resolution issues due to our docker-
compose setup, so we'll try and trap those, and log out appropriately

Lastly we've found (via the kafka exception) that if unknown exceptions
are thrown during jaxrs methods, that we're not logging them because our
exception handler was silently discarding all the useful debug info, so
lets log that out too.

Signed-off-by: Ozzy Osborne <ozzy@ca.ibm.com>